feat: rename togglePoint to addDecimal and change its behaviour
