+++ src/functions/addDecimal/utils.ts
@@ -1,0 +1,9 @@
+import { CalculationString } from '@src/types';
+
+const isEndsWithDecimal = (calculationString: CalculationString) => {
+  const decimalRegex = /^.*(\D{1})?\d+\.\d+$/;
+
+  return decimalRegex.test(calculationString);
+};
+
+export { isEndsWithDecimal };
