+++ src/functions/index.ts
@@ -1,7 +1,7 @@
 export * from './getResult';
 export * from './addNumber';
 export * from './addOperator';
-export * from './togglePoint';
+export * from './addDecimal';
 export * from './clearAll';
 export * from './clearOne';
 export * from './addThousandsSeparator';
